+++ r-naming.tf
@@ -1,5 +1,15 @@
 resource "azurecaf_name" "vm" {
   name          = var.stack
+  resource_type = "azurerm_linux_virtual_machine"
+  prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "vm"])
+  use_slug      = var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+resource "azurecaf_name" "hostname" {
+  name          = var.stack
   resource_type = "azurerm_windows_virtual_machine"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
   suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "vm"])